requirements.txt
mkdocs
mkdocs-material>=8.5.3
pillow
